Refactor: Server for multiple clients
